# api/wellness-buddy-openapi.yaml
openapi: 3.0.3
info:
  title: Wellness Buddy API
  version: 1.0.0
  description: |
    REST API for the Wellness Buddy app. The client must never access the
    database directly; all reads and writes go through this API layer.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /goals:
    get:
      tags: [Goals]
      summary: List goals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: category
          schema:
            type: string
            description: Domain of the goal (health, finance, legal, lifestyle).
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Goal'
    post:
      tags: [Goals]
      summary: Create a goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalCreate'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
  /goals/{goalId}:
    get:
      tags: [Goals]
      summary: Get a goal by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Goals]
      summary: Update a goal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalUpdate'
      responses:
        '200':
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Goals]
      summary: Delete a goal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Goal deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /journal-entries:
    get:
      tags: [Journal]
      summary: List journal entries
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: goalId
          schema:
            type: string
            format: uuid
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JournalEntry'
    post:
      tags: [Journal]
      summary: Create a journal entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryCreate'
      responses:
        '201':
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
  /journal-entries/{entryId}:
    get:
      tags: [Journal]
      summary: Get a journal entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EntryId'
      responses:
        '200':
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Journal]
      summary: Update a journal entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EntryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryUpdate'
      responses:
        '200':
          description: Updated journal entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Journal]
      summary: Delete a journal entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EntryId'
      responses:
        '204':
          description: Journal entry deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /memberships:
    get:
      tags: [Memberships]
      summary: List memberships
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: clubId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Membership'
    post:
      tags: [Memberships]
      summary: Join a club
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '409':
          description: User already a member of this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /memberships/{membershipId}:
    delete:
      tags: [Memberships]
      summary: Leave a club
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MembershipId'
      responses:
        '204':
          description: Membership deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /posts:
    get:
      tags: [Posts]
      summary: List posts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      tags: [Posts]
      summary: Create a post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /posts/{postId}:
    patch:
      tags: [Posts]
      summary: Update a post (e.g., approve/reject)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Posts]
      summary: Delete a post
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    GoalId:
      name: goalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Goal:
      type: object
      properties:
        goal_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        category:
          type: string
          description: Domain (health, finance, legal, lifestyle)
        description:
          type: string
        target_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GoalCreate:
      type: object
      required: [user_id, category, description]
      properties:
        user_id:
          type: string
          format: uuid
        category:
          type: string
        description:
          type: string
        target_date:
          type: string
          format: date
    GoalUpdate:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        target_date:
          type: string
          format: date
    JournalEntry:
      type: object
      properties:
        entry_id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        entry_date:
          type: string
          format: date
        content:
          type: string
        mood_rating:
          type: integer
          minimum: 1
          maximum: 5
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JournalEntryCreate:
      type: object
      required: [goal_id, user_id, entry_date]
      properties:
        goal_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        entry_date:
          type: string
          format: date
        content:
          type: string
        mood_rating:
          type: integer
          minimum: 1
          maximum: 5
    JournalEntryUpdate:
      type: object
      properties:
        entry_date:
          type: string
          format: date
        content:
          type: string
        mood_rating:
          type: integer
          minimum: 1
          maximum: 5
...
    Membership:
      type: object
      properties:
        membership_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        club_id:
          type: string
          format: uuid
        joined_at:
          type: string
          format: date-time
    MembershipCreate:
      type: object
      required: [user_id, club_id]
      properties:
        user_id:
          type: string
          format: uuid
        club_id:
          type: string
          format: uuid
...
    Post:
      type: object
      properties:
        post_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [pending, approved, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PostCreate:
      type: object
      required: [user_id, title, rating]
      properties:
        user_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [pending, approved, rejected]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true